+++ src/Sentry/Protocol/Envelopes/Envelope.cs
@@ -17,10 +17,6 @@ namespace Sentry.Protocol.Envelopes
     /// </summary>
     public sealed class Envelope : ISerializable, IDisposable
     {
-        private const string SdkKey = "sdk";
-        private const string EventIdKey = "event_id";
-        private const string SentAtKey = "sent_at";
-
         /// <summary>
         /// Header associated with the envelope.
         /// </summary>
@@ -44,7 +40,7 @@ namespace Sentry.Protocol.Envelopes
         /// Attempts to extract the value of "event_id" header if it's present.
         /// </summary>
         public SentryId? TryGetEventId() =>
-            Header.TryGetValue(EventIdKey, out var value) &&
+            Header.TryGetValue("event_id", out var value) &&
             value is string valueString &&
             Guid.TryParse(valueString, out var guid)
                 ? new SentryId(guid)
@@ -58,7 +54,7 @@ namespace Sentry.Protocol.Envelopes
         {
             // Append the sent_at header, except when writing to disk
             var headerItems = !stream.IsFileStream()
-                ? Header.Append(SentAtKey, clock.GetUtcNow())
+                ? Header.Append("sent_at", clock.GetUtcNow())
                 : Header;
 
             var writer = new Utf8JsonWriter(stream);
@@ -78,7 +74,7 @@ namespace Sentry.Protocol.Envelopes
         {
             // Append the sent_at header, except when writing to disk
             var headerItems = !stream.IsFileStream()
-                ? Header.Append(SentAtKey, clock.GetUtcNow())
+                ? Header.Append("sent_at", clock.GetUtcNow())
                 : Header;
 
             using var writer = new Utf8JsonWriter(stream);
@@ -133,30 +129,25 @@ namespace Sentry.Protocol.Envelopes
         public void Dispose() => Items.DisposeAll();
 
         // limited SDK information (no packages)
-        private static readonly IReadOnlyDictionary<string, string?> SdkHeader = new Dictionary<string, string?>(2, StringComparer.Ordinal)
-        {
-            ["name"] = SdkVersion.Instance.Name,
-            ["version"] = SdkVersion.Instance.Version
-        };
-
-        private static readonly IReadOnlyDictionary<string, object?> DefaultHeader = new Dictionary<string, object?>(1, StringComparer.Ordinal)
-        {
-            ["sdk"] = SdkHeader
-        };
-
-        private static IReadOnlyDictionary<string, object?> CreateHeader(SentryId? eventId = null)
-        {
-            if (eventId is null)
+        private static readonly IReadOnlyDictionary<string, string?> SdkHeader =
+            new Dictionary<string, string?>(2, StringComparer.Ordinal)
             {
-                return DefaultHeader;
-            }
+                ["name"] = SdkVersion.Instance.Name,
+                ["version"] = SdkVersion.Instance.Version
+            }.AsReadOnly();
 
-            return new Dictionary<string, object?>(2, StringComparer.Ordinal)
+        private static readonly IReadOnlyDictionary<string, object?> DefaultHeader =
+            new Dictionary<string, object?>(1, StringComparer.Ordinal)
             {
-                [SdkKey] = SdkHeader,
-                [EventIdKey] = eventId.Value.ToString()
+                ["sdk"] = SdkHeader
+            }.AsReadOnly();
+
+        private static Dictionary<string, object?> CreateHeader(SentryId eventId, int extraCapacity = 0) =>
+            new(2 + extraCapacity, StringComparer.Ordinal)
+            {
+                ["sdk"] = SdkHeader,
+                ["event_id"] = eventId.ToString()
             };
-        }
 
         /// <summary>
         /// Creates an envelope that contains a single event.
@@ -233,7 +224,16 @@ namespace Sentry.Protocol.Envelopes
         /// </summary>
         public static Envelope FromTransaction(Transaction transaction)
         {
-            var header = CreateHeader(transaction.EventId);
+            Dictionary<string, object?> header;
+            if (transaction.DynamicSamplingContext is { } dsc)
+            {
+                header = CreateHeader(transaction.EventId, extraCapacity: 1);
+                header["trace"] = dsc.Items;
+            }
+            else
+            {
+                header = CreateHeader(transaction.EventId);
+            }
 
             var items = new[]
             {
@@ -248,7 +248,7 @@ namespace Sentry.Protocol.Envelopes
         /// </summary>
         public static Envelope FromSession(SessionUpdate sessionUpdate)
         {
-            var header = CreateHeader();
+            var header = DefaultHeader;
 
             var items = new[]
             {
@@ -263,7 +263,7 @@ namespace Sentry.Protocol.Envelopes
         /// </summary>
         internal static Envelope FromClientReport(ClientReport clientReport)
         {
-            var header = CreateHeader();
+            var header = DefaultHeader;
 
             var items = new[]
             {
@@ -297,7 +297,7 @@ namespace Sentry.Protocol.Envelopes
                 ?? throw new InvalidOperationException("Envelope header is malformed.");
 
             // The sent_at header should not be included in the result
-            header.Remove(SentAtKey);
+            header.Remove("sent_at");
 
             return header;
         }